lading page + announcement v1
backend:
store announcementu,
announcement controller, middleware pre checkrole pre admin / garant
laravel password pre API autentifikáciu

frontend:
pinia store na správu stavu
quill.js editor
zobrazenie oznámení
role-based UI v dashboard view
jednotný navbar v app.vue
prepísanie tailwind na bootstrap classes + odstránenie tailwind komponentov
fix yellow warnings

+++ backend/database/migrations/2025_10_11_134950_add_foreign_keys_to_companies_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -12,9 +13,26 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('companies', function (Blueprint $table) {
-            $table->foreign(['address_id'], 'companies_address_FK')->references(['id'])->on('address')->onUpdate('cascade')->onDelete('cascade');
-            $table->foreign(['user_id'], 'companies_user_FK')->references(['id'])->on('users')->onUpdate('cascade')->onDelete('cascade');
+            if (!$this->foreignKeyExists('companies', 'companies_address_FK')) {
+                $table->foreign(['address_id'], 'companies_address_FK')->references(['id'])->on('address')->onUpdate('cascade')->onDelete('cascade');
+            }
+            if (!$this->foreignKeyExists('companies', 'companies_user_FK')) {
+                $table->foreign(['user_id'], 'companies_user_FK')->references(['id'])->on('users')->onUpdate('cascade')->onDelete('cascade');
+            }
         });
+    }
+
+    private function foreignKeyExists($table, $constraintName)
+    {
+        $foreignKeys = DB::select("
+            SELECT CONSTRAINT_NAME 
+            FROM information_schema.KEY_COLUMN_USAGE 
+            WHERE TABLE_SCHEMA = DATABASE() 
+            AND TABLE_NAME = ? 
+            AND CONSTRAINT_NAME = ?
+        ", [$table, $constraintName]);
+        
+        return !empty($foreignKeys);
     }
 
     /**
